allowing to query topics from kafka, instead of always querying from zookeeper, and refactoring
