+++ kafka_utils/kafka_consumer_manager/commands/offset_get.py
@@ -83,7 +83,11 @@ class OffsetGet(OffsetManagerBase):
         client.load_metadata_for_topics()
 
         topics_dict = cls.preprocess_args(
-            args.groupid, args.topic, args.partitions, cluster_config, client,
+            groupid=args.groupid,
+            topic=args.topic,
+            partitions=args.partitions,
+            cluster_config=cluster_config,
+            client=client,
             quiet=args.json
         )
