+++ kafka_utils/kafka_consumer_manager/commands/offset_save.py
@@ -80,11 +80,11 @@ class OffsetSave(OffsetManagerBase):
         client.load_metadata_for_topics()
 
         topics_dict = cls.preprocess_args(
-            args.groupid,
-            args.topic,
-            args.partitions,
-            cluster_config,
-            client,
+            groupid=args.groupid,
+            topic=args.topic,
+            partitions=args.partitions,
+            cluster_config=cluster_config,
+            client=client,
         )
         try:
             consumer_offsets_metadata = get_consumer_offsets_metadata(
