+++ kafka_utils/kafka_consumer_manager/util.py
@@ -16,10 +16,23 @@ from __future__ import absolute_import
 from __future__ import print_function
 from __future__ import unicode_literals
 
+import logging
 import sys
 from collections import defaultdict
 
+from kafka.common import ConsumerTimeout
+from kafka.common import FailedPayloadsError
+from kafka.common import KafkaUnavailableError
+from kafka.common import LeaderNotAvailableError
+from kafka.common import NotLeaderForPartitionError
+from kafka.consumer import KafkaConsumer
+from kafka.util import read_short_string
+from kafka.util import relative_unpack
 from kazoo.exceptions import NodeExistsError
+
+from kafka_utils.util.offsets import get_topics_watermarks
+
+CONSUMER_OFFSET_TOPIC = '__consumer_offsets'
 
 
 def preprocess_topics(source_groupid, source_topics, dest_groupid, topics_dest_group):
@@ -113,3 +126,91 @@ def prompt_user_input(in_str):
             sys.exit(0)
         if answer == "y" or answer == "yes":
             return
+
+
+class InvalidMessageException(Exception):
+    pass
+
+
+class KafkaGroupReader:
+
+    def __init__(self, kafka_config):
+        self.log = logging.getLogger(__name__)
+        self.kafka_config = kafka_config
+        self.kafka_groups = defaultdict(set)
+        self.finished_partitions = set()
+
+    def read_groups(self):
+        self.log.info("Kafka consumer running")
+        self.consumer = KafkaConsumer(
+            CONSUMER_OFFSET_TOPIC,
+            group_id='offset_monitoring_consumer',
+            bootstrap_servers=self.kafka_config.broker_list,
+            auto_offset_reset='smallest',
+            auto_commit_enable=False,
+            consumer_timeout_ms=10000,
+        )
+        self.log.info("Consumer ready")
+        self.watermarks = self.get_current_watermarks()
+        while not self.finished():
+            try:
+                message = self.consumer.next()
+                max_offset = self.get_max_offset(message.partition)
+                if message.offset >= max_offset - 1:
+                    self.finished_partitions.add(message.partition)
+            except ConsumerTimeout:
+                break
+            except (
+                    FailedPayloadsError,
+                    KafkaUnavailableError,
+                    LeaderNotAvailableError,
+                    NotLeaderForPartitionError,
+            ) as e:
+                self.log.warning("Got %s, retrying", e.__class__.__name__)
+            self.process_consumer_offset_message(message)
+        return self.kafka_groups
+
+    def parse_consumer_offset_message(self, message):
+        key = bytearray(message.key)
+        ((key_schema,), cur) = relative_unpack('>h', key, 0)
+        if key_schema not in [0, 1]:
+            raise InvalidMessageException()   # This is not an offset commit message
+        (group, cur) = read_short_string(key, cur)
+        (topic, cur) = read_short_string(key, cur)
+        ((partition,), cur) = relative_unpack('>l', key, cur)
+        if message.value:
+            value = bytearray(message.value)
+            ((value_schema,), cur) = relative_unpack('>h', value, 0)
+            if value_schema not in [0, 1]:
+                raise InvalidMessageException()  # Unrecognized message value
+            ((offset,), cur) = relative_unpack('>q', value, cur)
+        else:
+            offset = None  # Offset was deleted
+        return str(group), str(topic), partition, offset
+
+    def process_consumer_offset_message(self, message):
+        try:
+            group, topic, partition, offset = self.parse_consumer_offset_message(message)
+        except InvalidMessageException:
+            return
+
+        if offset:
+            self.kafka_groups[group].add(topic)
+        else:  # No offset means group deletion
+            self.kafka_groups.pop(group, None)
+
+    def get_current_watermarks(self):
+        self.consumer._client.load_metadata_for_topics()
+        offsets = get_topics_watermarks(
+            self.consumer._client,
+            [CONSUMER_OFFSET_TOPIC],
+        )
+        return {partition: offset for partition, offset
+                in offsets[CONSUMER_OFFSET_TOPIC].iteritems()
+                if offset.highmark > offset.lowmark}
+
+    def get_max_offset(self, partition):
+        return self.watermarks[partition].highmark
+
+    def finished(self):
+        return len(self.finished_partitions) >= len(self.watermarks)
